move profile property value logic to op
